+++ src/Packer.ts
@@ -1,11 +1,10 @@
 import {
     compressScriptFiles,
     RubyCompressScriptService,
-    RubyScriptService,
 } from "./commands/ExtractScriptFiles";
 import { ConfigService } from "./services/ConfigService";
 import { LoggingService } from "./services/LoggingService";
-import { TARGET_SCRIPT_FILE_NAME, Unpacker } from "./Unpacker";
+import { Unpacker } from "./Unpacker";
 import * as path from "path";
 import { Path } from "./utils/Path";